requirements.txt
flask>=2.0
requests>=2.28
transformers>=4.0
torch
Pillow
